Address some request changes
